+++ packages/core/src/matterbridgeEndpoint.ts
@@ -3388,7 +3388,7 @@ export class MatterbridgeEndpoint extends Endpoint {
    * All operating modes NOT supported by a lock SHALL be set to one. The value of the OperatingMode enumeration defines the related bit to be set.
    */
   createDefaultDoorLockClusterServer(lockState: DoorLock.LockState = DoorLock.LockState.Locked, lockType: DoorLock.LockType = DoorLock.LockType.DeadBolt): this {
-    this.behaviors.require(MatterbridgeDoorLockServer.enable({ events: { doorLockAlarm: true, lockOperation: true, lockOperationError: true } }), {
+    this.behaviors.require(MatterbridgeDoorLockServer.with().enable({ events: { doorLockAlarm: true, lockOperation: true, lockOperationError: true } }), {
       lockState,
       lockType,
       /** This attribute SHALL indicate if the lock is currently able to (Enabled) or not able to (Disabled) process remote Lock, Unlock, or Unlock with Timeout commands. */
@@ -3412,6 +3412,66 @@ export class MatterbridgeEndpoint extends Endpoint {
       supportedOperatingModes: { normal: false, vacation: true, privacy: true, noRemoteLockUnlock: false, passage: true, alwaysSet: 2047 },
       autoRelockTime: 0, // 0=disabled
     });
+    return this;
+  }
+
+  /**
+   * Creates a default PIN door lock cluster server.
+   *
+   * @param {DoorLock.LockState} [lockState] - The initial state of the lock (default: Locked).
+   * @param {DoorLock.LockType} [lockType] - The type of the lock (default: DeadBolt).
+   * @returns {this} The current MatterbridgeEndpoint instance for chaining.
+   *
+   * @remarks
+   * All operating modes NOT supported by a lock SHALL be set to one. The value of the OperatingMode enumeration defines the related bit to be set.
+   */
+  createPinDoorLockClusterServer(lockState: DoorLock.LockState = DoorLock.LockState.Locked, lockType: DoorLock.LockType = DoorLock.LockType.DeadBolt): this {
+    this.behaviors.require(
+      MatterbridgeDoorLockServer.with(DoorLock.Feature.PinCredential, DoorLock.Feature.CredentialOverTheAirAccess).enable({
+        events: { doorLockAlarm: true, lockOperation: true, lockOperationError: true },
+      }),
+      {
+        lockState,
+        lockType,
+        /** This attribute SHALL indicate if the lock is currently able to (Enabled) or not able to (Disabled) process remote Lock, Unlock, or Unlock with Timeout commands. */
+        actuatorEnabled: true,
+        /** This attribute SHALL indicate the current operating mode of the lock as defined in OperatingModeEnum */
+        operatingMode: DoorLock.OperatingMode.Normal,
+        /**
+         * This attribute SHALL contain a bitmap with all operating bits of the OperatingMode attribute supported
+         * by the lock. All operating modes NOT supported by a lock SHALL be set to one. The value of
+         * the OperatingMode enumeration defines the related bit to be set.
+         * OperatingModesBitmap.Normal and OperatingModesBitmap.noRemoteLockUnlock are mandatory and SHALL always be supported.
+         * Default value 0xFFF6 (1111 1111 1111 0110) means:
+         * - normal: false (bit 0)
+         * - vacation: true (bit 1)
+         * - privacy: true (bit 2)
+         * - noRemoteLockUnlock: false (bit 3)
+         * - passage: true (bit 4)
+         * Special case of inverted bitmap: add also alwaysSet = 2047 (0000 0111 1111 1111) to have all bits set except the unsupported ones.
+         * Specs: "Any bit that is not yet defined in OperatingModesBitmap SHALL be set to 1."
+         */
+        supportedOperatingModes: { normal: false, vacation: true, privacy: true, noRemoteLockUnlock: false, passage: true, alwaysSet: 2047 },
+        autoRelockTime: 0, // 0=disabled
+        // PinCredential feature attributes
+        numberOfPinUsersSupported: 10,
+        minPinCodeLength: 4,
+        maxPinCodeLength: 10,
+        // PinCredential feature attributes (Not User feature)
+        sendPinOverTheAir: true,
+        // PinCredential and CredentialOverTheAirAccess features attributes
+        requirePinForRemoteOperation: true,
+        // PinCredential or RfidCredential feature attributes
+        wrongCodeEntryLimit: 5,
+        userCodeTemporaryDisableTime: 60,
+        // User feature attributes
+        /*
+        numberOfTotalUsersSupported: 10,
+        credentialRulesSupport: { single: true },
+        numberOfCredentialsSupportedPerUser: 1,
+        */
+      },
+    );
     return this;
   }